// src/app/farmer-beneficiary-management/models/models/farmer-beneficiary.model.ts
export interface FarmerBeneficiary {
    isSelected: boolean;
    state: string;
    district: string;
    taluka: string;
    village: string;
    farmerName: string;
    beneficiaryId: string;
    shopNo: string;
    rcNo: string;
    uniqueNoOld: string;
    aadhaarNumber: string;
    aadhaarReferenceNo: string;
    firstName: string;
    middleName: string;
    surname: string;
    fullName: String;
    bankName: string;
    accountNumber: string;
    ifscCode: string;   
    beneficiaryName: string;
    totalMembers: number;
    amount: number;
    srNo: number;

}